% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_sba.R
\name{Fz_fun}
\alias{Fz_fun}
\title{Compute the latent data CDF}
\usage{
Fz_fun(z, weights = NULL, mean_vec = NULL, sd_vec)
}
\arguments{
\item{z}{vector of points at which the CDF of z is evaluated}

\item{weights}{\code{n}-dimensional vector of weights; if NULL,
assume 1/n}

\item{mean_vec}{\code{n}-dimensional vector of means; if NULL,
assume mean zero}

\item{sd_vec}{\code{n}-dimensional vector of standard deviations}
}
\value{
CDF of z evaluated at \code{z}
}
\description{
Assuming a Gaussian latent data distribution (given x),
compute the CDF on a grid of points
}
