% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_sba.R
\name{g_fun}
\alias{g_fun}
\title{Compute the transformation}
\usage{
g_fun(y, Fy_eval, z, Fz_eval)
}
\arguments{
\item{y}{vector of points at which the CDF of y is evaluated}

\item{Fy_eval}{CDF of y evaluated at \code{y}}

\item{z}{vector of points at which the CDF of z is evaluated}

\item{Fz_eval}{CDF of z evaluated at \code{z}}
}
\value{
A smooth monotone function which can be used for evaluations of the transformation.
}
\description{
Given the CDFs of z and y, compute a smoothed function
to evaluate the transformation
}
