% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_sba.R
\name{g_inv_approx}
\alias{g_inv_approx}
\title{Approximate inverse transformation}
\usage{
g_inv_approx(g, t_grid)
}
\arguments{
\item{g}{the transformation function}

\item{t_grid}{grid of arguments at which to evaluate the transformation function}
}
\value{
A function which can be used for evaluations of the
(approximate) inverse transformation function.
}
\description{
Compute the inverse function of a transformation \code{g} based on a grid search.
}
