% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_varsel.R
\name{square_stabilize}
\alias{square_stabilize}
\title{Numerically stabilize the squared elements}
\usage{
square_stabilize(vec)
}
\arguments{
\item{vec}{vector of inputs to be squared}
}
\value{
a vector of the same length as `vec`
}
\description{
Given a vector to be squared, add a numeric buffer
for the elements very close to zero.
}
