% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_varsel.R
\name{SSR_gprior}
\alias{SSR_gprior}
\title{Compute the sum-squared-residuals term under Zellner's g-prior}
\usage{
SSR_gprior(y, X = NULL, psi)
}
\arguments{
\item{y}{vector of response variables}

\item{X}{matrix of covariates; if NULL, return \code{sum(y^2)}}

\item{psi}{prior variance (g-prior)}
}
\value{
a positive scalar
}
\description{
These sum-squared-residuals (SSR) arise in the variance (or precision)
term under 1) Zellner's g-prior on the coefficients and a Gamma prior on the
error precision and 2) marginalization over the coefficients.
}
